[IMP] shopinvader_router_helper: Use odoo_test_helper for test virtual models
